+++ lustre/mdc/mdc_internal.h
@@ -22,14 +22,14 @@
 #define MDC_INTERNAL_H
 
 void mdc_getattr_pack(struct lustre_msg *msg, int offset,
-                      int valid, int flags, struct mdc_op_data *data);
+                      __u64 valid, int flags, struct mdc_op_data *data);
 void mdc_open_pack(struct lustre_msg *msg, int offset,
                    struct mdc_op_data *op_data, __u32 mode, __u64 rdev,
                    __u32 flags, const void *lmm, int lmmlen);
 void mdc_readdir_pack(struct ptlrpc_request *req, int req_offset,
                       __u64 offset, __u32 size, struct lustre_id *mdc_id);
 void mdc_close_pack(struct ptlrpc_request *req, int offset, struct obdo *oa,
-		    int valid, struct obd_client_handle *och);
+		    __u64 valid, struct obd_client_handle *och);
 
 struct mdc_open_data {
         struct obd_client_handle *mod_och;
